refactor: transparentable header on project page
