+++ lib/global_utils.ts
@@ -1,0 +1,2 @@
+export const isBrowser = (): boolean => !isServerSide();
+export const isServerSide = (): boolean => typeof window === 'undefined';
